FIIIXXXXX

+++ src/main/java/game/entities/targeting/Arrow.java
@@ -9,7 +9,7 @@ import game.scenes.GameScreen;
 
 public class Arrow extends Projectile implements Collided, SceneBorderCrossingWatcher {
 
-    public Arrow(Coordinate2D initialLocation, int shootAngle, GameScreen gameScreen, Tower shootingTower) {
+    public Arrow(Coordinate2D initialLocation, double shootAngle, GameScreen gameScreen, Tower shootingTower) {
         super("sprites/arrow.png", initialLocation, gameScreen, shootingTower);
         int movementSpeed = 25;
         setAnchorPoint(AnchorPoint.CENTER_CENTER);
